+++ e2e/testsuite/testsuite.go
@@ -23,6 +23,7 @@ import (
 	"google.golang.org/grpc/credentials/insecure"
 
 	"github.com/cosmos/ibc-go/e2e/testconfig"
+	"github.com/cosmos/ibc-go/e2e/testvalues"
 	feetypes "github.com/cosmos/ibc-go/v5/modules/apps/29-fee/types"
 	clienttypes "github.com/cosmos/ibc-go/v5/modules/core/02-client/types"
 	channeltypes "github.com/cosmos/ibc-go/v5/modules/core/04-channel/types"
@@ -50,6 +51,9 @@ type E2ETestSuite struct {
 	DockerClient   *dockerclient.Client
 	network        string
 	startRelayerFn func(relayer ibc.Relayer)
+
+	// pathNameIndex is the latest index to be used for generating paths
+	pathNameIndex uint64
 }
 
 // GRPCClients holds a reference to any GRPC clients that are needed by the tests.
@@ -110,8 +114,7 @@ func (s *E2ETestSuite) SetupChainsRelayerAndChannel(ctx context.Context, channel
 
 	r := newCosmosRelayer(s.T(), testconfig.FromEnv(), s.logger, s.DockerClient, s.network)
 
-	pathName := fmt.Sprintf("%s-path", s.T().Name())
-	pathName = strings.ReplaceAll(pathName, "/", "-")
+	pathName := s.generatePathName()
 
 	ic := ibctest.NewInterchain().
 		AddChain(chainA).
@@ -157,6 +160,39 @@ func (s *E2ETestSuite) SetupChainsRelayerAndChannel(ctx context.Context, channel
 	chainAChannels, err := r.GetChannels(ctx, eRep, chainA.Config().ChainID)
 	s.Require().NoError(err)
 	return r, chainAChannels[len(chainAChannels)-1]
+}
+
+// generatePathName generates the path name using the test suites name
+func (s *E2ETestSuite) generatePathName() string {
+	pathName := fmt.Sprintf("%s-path-%d", s.T().Name(), s.pathNameIndex)
+	s.pathNameIndex++
+	return strings.ReplaceAll(pathName, "/", "-")
+}
+
+// generatePath generates the path name using the test suites name
+func (s *E2ETestSuite) generatePath(ctx context.Context, relayer ibc.Relayer) string {
+	chainA, chainB := s.GetChains()
+	chainAID := chainA.Config().ChainID
+	chainBID := chainB.Config().ChainID
+
+	pathName := s.generatePathName()
+	err := relayer.GeneratePath(ctx, s.GetRelayerExecReporter(), chainAID, chainBID, pathName)
+	s.Require().NoError(err)
+
+	return pathName
+}
+
+// SetupClients creates clients on chainA and chainB using the provided create client options
+func (s *E2ETestSuite) SetupClients(ctx context.Context, relayer ibc.Relayer, opts ibc.CreateClientOptions) {
+	pathName := s.generatePath(ctx, relayer)
+	err := relayer.CreateClients(ctx, s.GetRelayerExecReporter(), pathName, opts)
+	s.Require().NoError(err)
+}
+
+// UpdateClients updates clients on chainA and chainB
+func (s *E2ETestSuite) UpdateClients(ctx context.Context, relayer ibc.Relayer, pathName string) {
+	err := relayer.UpdateClients(ctx, s.GetRelayerExecReporter(), pathName)
+	s.Require().NoError(err)
 }
 
 // GetChains returns two chains that can be used in a test. The pair returned
@@ -356,26 +392,11 @@ func (s *E2ETestSuite) createCosmosChains(chainOptions testconfig.ChainOptions) 
 
 	logger := zaptest.NewLogger(s.T())
 
-	numValidators, numFullNodes := getValidatorsAndFullNodes(chainOptions)
+	numValidators, numFullNodes := 4, 1
 
 	chainA := cosmos.NewCosmosChain(s.T().Name(), *chainOptions.ChainAConfig, numValidators, numFullNodes, logger)
 	chainB := cosmos.NewCosmosChain(s.T().Name(), *chainOptions.ChainBConfig, numValidators, numFullNodes, logger)
 	return chainA, chainB
-}
-
-// getValidatorsAndFullNodes returns the number of validators and full nodes which should be used
-// for the given chain config.
-func getValidatorsAndFullNodes(chainOptions testconfig.ChainOptions) (int, int) {
-	// TODO: the icad tests are failing with a larger number of validators.
-	// this function can be removed once https://github.com/cosmos/ibc-go/issues/2104 is resolved.
-	numValidators := 4
-	numFullNodes := 1
-	isIcadImage := strings.Contains(chainOptions.ChainAConfig.Images[0].Repository, "icad")
-	if isIcadImage {
-		numValidators = 1
-		numFullNodes = 0
-	}
-	return numValidators, numFullNodes
 }
 
 // GetRelayerExecReporter returns a testreporter.RelayerExecReporter instances
@@ -430,7 +451,7 @@ func (s *E2ETestSuite) ExecuteGovProposal(ctx context.Context, chain *cosmos.Cos
 	s.Require().NoError(err)
 	s.Require().Equal(govtypes.StatusVotingPeriod, proposal.Status)
 
-	time.Sleep(time.Second * 30) // pass proposal
+	time.Sleep(testvalues.VotingPeriod) // pass proposal
 
 	proposal, err = s.QueryProposal(ctx, chain, 1)
 	s.Require().NoError(err)